#144: extract general description row and add it to the submission modal
