+++ src/components/conference/NewConference/Page/PerformanceModal/Form.svelte
@@ -1,11 +1,13 @@
 <script>
     import NameAndAuthorRow from "./NameAndAuthorRow.svelte";
     import TimeLocationTypeRows from "./TimeLocationTypeRows.svelte";
+    import Description from "./Description.svelte";
 </script>
 
 <form>
     <NameAndAuthorRow />
     <TimeLocationTypeRows />
+    <Description />
 </form>
 
 <style>
